Combine Visibility Display and Select. i18n for both
