+++ src/components/ProfilePage/elements/Visibility.messages.jsx
@@ -1,13 +1,13 @@
 import { defineMessages } from 'react-intl';
 
 const messages = defineMessages({
-  'profile.visibilityselect.who.just.me': {
-    id: 'profile.visibilityselect.who.just.me',
+  'profile.visibility.who.just.me': {
+    id: 'profile.visibility.who.just.me',
     defaultMessage: 'Just me',
     description: 'What users can see this area?',
   },
-  'profile.visibilityselect.who.everyone': {
-    id: 'profile.visibilityselect.who.everyone',
+  'profile.visibility.who.everyone': {
+    id: 'profile.visibility.who.everyone',
     defaultMessage: 'Everyone on edX',
     description: 'What users can see this area?',
   },
